Add patches for Intel TXT support
Set of patches which adds TPM1.2 support for Intel TXT in GRUB2 for
TrenchBoot.

This is necessary to create Proof of Concept for TrenchBoot Anti Evil
Maid for QubesOS.

The TrenchBoot support hasn't been implemented and verified with TPM 1.2
on Intel TXT path. This changes ensures the TPM 1.2 is also supported
for older Intel hardware with Intel TXT.

Signed-off-by: Tomasz Żyjewski <tomasz.zyjewski@3mdeb.com>