Add another fallback boot pathway for UEFI: default image
In addition to the vendor image location (e.g. \EFI\qubes\grubx64.efi)
which is used for the Qubes OS boot entry, install a copy of the file to
the architecture's default image location (e.g. \EFI\BOOT\BOOTX64.efi)
unless that's already occupied.

If the boot entry is missing, the system can discover the default image.
Compared to the startup.nsh hack, booting a default image works even if
a UEFI Shell is not being booted (or if it's not looking for startup.nsh
on our drive); OTOH it may require the user to manually select the drive
in a menu.

Making such a file copy is also the approach taken by systemd's 'bootctl
install', except that we avoid clobbering an existing file. (The
%config(noreplace) mechanism that's used here isn't ideal in case e.g.
grub2-efi-x64 is ever uninstalled, because that would move the existing
file to a .rpmsave suffixed name instead of leaving it alone. But
/etc/dnf/protected.d/ already marks the package as something not meant
to be uninstalled.)

Fixes QubesOS/qubes-issues#8363